ContentView.swift different per platform
